fix(decimal-precision): fix BusinessCase.estimated_cost and frontend handling
Backend:
- Changed BusinessCase.estimated_cost from Float to Numeric(10, 2)
- Added Decimal type with 2dp rounding validator in schema

Frontend (business-cases.vue):
- Changed estimated_cost interface from number to string
- Updated form handling to use string-based values
- Updated template to use parseFloat() for display formatting

Status: Partial fix - 6 more frontend pages need updating

+++ backend/app/models.py
@@ -105,7 +105,7 @@ class BusinessCase(Base):
     requestor = Column(String(255))
     dept = Column(String(255))
     lead_group_id = Column(Integer, ForeignKey("user_group.id"))
-    estimated_cost = Column(Float)
+    estimated_cost = Column(Numeric(10, 2))
     status = Column(String(50))
 
     # Audit
